daemon: Don't delete files from the store with guix build -K --rounds=3 --check.
Specifying all three of -K, --check, --rounds=3 caused valid packages to be
deleted from the store.

When `buildMode == bmCheck`, the built package is not renamed from
`actualPath` to `i.second.path.  Then later, when `settings.keepFailed` and
`curRound < nrRounds`, `i.second.path` was renamed to `i.second.path + "-check"`
anyway.

* nix/libstore/build.cc (DerivationGoal::registerOutputs): Add check for bmCheck
before renaming store path.
* tests/store.scm ("build-things, check mode"): Ensure file is a valid path.
("build-things, check mode + keep-failed"): New test.

Fixes: #9631.
Merges: #9632

+++ nix/libstore/build.cc
@@ -3370,7 +3370,10 @@ void DerivationGoal::registerOutputs()
         assert(false); // shouldn't happen
     }
 
-    if (settings.keepFailed) {
+    /* Rename the built package when doing multiple rounds, but
+       don't move the package when running in check mode, since
+       that would remove the valid package from the store. */
+    if (settings.keepFailed && buildMode != bmCheck) {
         for (auto & i : drv.outputs) {
             Path prev = i.second.path + checkSuffix;
             if (pathExists(prev)) deletePath(prev);
